Menage danbs Qt. Modif vol and measAreas type (size_t). Ajout labelWithArea et areaOpen.

+++ Base/include/private/DLabelMeasures.hpp
@@ -46,12 +46,12 @@ namespace smil
 
     /**
     * Measure label areas.
-    * Return a map(labelValue, double) with the area of each label value.
+    * Return a map(labelValue, size_t) with the area of each label value.
     */
     template <class T>
-    map<T, double> measAreas(Image<T> &imIn)
+    map<T, size_t> measAreas(Image<T> &imIn)
     {
-	map<T, double> area;
+	map<T, size_t> area;
 
 	ASSERT(CHECK_ALLOCATED(&imIn), RES_ERR_BAD_ALLOCATION, area);
